Refactor

+++ Project/Source/Forms/Boxes/StatisticsForm.cs
@@ -119,7 +119,7 @@ namespace Ordisoftware.HebrewWords
 
     private void InitializeMiddle()
     {
-      int lcount = (int)Math.Truncate((double)CountersTorah.CountLetters / 2.0);
+      int lcount = (int)Math.Truncate(CountersTorah.CountLetters / 2.0);
       int index = 0;
       var books = from book in MainForm.Instance.DataSet.Books
                   where book.Number <= BooksBounds.Torah.Max
